Added Match and Bracket classes, Winner enum.
Match is a container that holds two wrestlers, their points,
and the matchWinner that is used to determine which wrestler
moves on to the next match. Will need to create a matchLoser
variable as well for Double Elimination brackets.

Bracket holds all the code to accept a list of Wrestlers and
separate them into Matches, while getting other matches
ready for the bracket. Some of that code needs to be
added still.

Winner is an enum used to keep track of whether the top or
bottom wrestler in the bracket won, and move them on
accordingly.