Changes to User creation, teamManage Activity and XML.
Coaches and Wrestlers now get a tournaments collection on creation so they can see the tournaments they are participating in.

Redid the constraints for the teamManage XML so that everything fits nicely.

Made changes to teamManage that makes the team selection spinner pull the list of team members from the database whenever it's changed. This list is automatically put into our ListView and displayed on screen.